detalles jugadores

+++ src/Interfaz/VentanaJugadores.java
@@ -215,7 +215,10 @@ public class VentanaJugadores extends javax.swing.JFrame implements Observer {
     private void jBtnBorrarActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtnBorrarActionPerformed
         DefaultTableModel model = (DefaultTableModel) tablaJugadores.getModel();
         if (tablaJugadores.isRowSelected(tablaJugadores.getSelectedRow())) {
-            miSistema.eliminarJugador(miSistema.getListaJugadores().get(tablaJugadores.getSelectedRow()));
+             if (JOptionPane.showConfirmDialog(rootPane, "¿Seguro/a que desea borrar el/la jugador/a?",
+                "Borrar Jugador/a", JOptionPane.YES_NO_OPTION) == JOptionPane.YES_OPTION)
+             miSistema.eliminarJugador(miSistema.getListaJugadores().get(tablaJugadores.getSelectedRow()));
+           
 //            model.removeRow(tabla.getSelectedRow());
         }else{
             if (tablaJugadores.getRowCount() == 0 || !tablaJugadores.isRowSelected(tablaJugadores.getSelectedRow())) {
